runtime: scope wide helpers to 32-bit targets
