+++ ssa/bounds_checks_test.go
@@ -44,6 +44,8 @@ func TestBoundsCheckModesIR(t *testing.T) {
 }
 
 func TestWideIndexBoundsCheck386(t *testing.T) {
+	t.Setenv("GOOS", "windows")
+	t.Setenv("GOARCH", "386")
 	prog := ssatest.NewProgram(t, &ssa.Target{GOOS: "windows", GOARCH: "386"})
 	t.Cleanup(prog.Dispose)
